backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
markdown==3.6
email-validator==2.1.0
passlib==1.7.4